Started PPU implementation

+++ App.cpp
@@ -85,7 +85,7 @@ void App::Run()
 		m_pRenderer->DrawFrame(m_Nes.m_Ppu.GetScreen().data());
 		m_pRenderer->RenderPresent();
 		std::chrono::duration<double> duration{ std::chrono::system_clock::now() - m_LastFrameTime };
-		m_ElapsedTime = duration.count();
+		m_ElapsedTime = (float)duration.count();
 	}
 }
